Initial ELF Loader
This pull request contains the initial ELF loader. It is
capable of loading multiple ELF files into memory,
relocating and resolving all symbols, and then executing
arbitrary entry points.

This pull request also contains a basic unit test framework
as well as a basic Makefile framework for compiling.

Tasks to be completed in the future include:
- Lazy loading for R_X86_64_JUMP_SLOT
- Using .hash instead of doing O(n) search
- Detecting duplicate symbols

[RFC]: https://groups.google.com/forum/#!topic/bareflank/8NO9AVGzlaU

Signed-off-by: Rian Quinn <quinnr@ainfosec.com>