Fix compilation error
The #define -> const char * conversion functions were causing
problems with compilations once we moved to #define to support
GCC. The following patch fixes an issue with the first attempt
to fix the compilation problems, as clang does not like the
keyword "class" being used. Instead, all of the functions have
been changed to use "value" which is consistent.

[ISSUE]: https://github.com/Bareflank/hypervisor/issues/10

Signed-off-by: Rian Quinn <quinnr@ainfosec.com>