Bareflank Manager
This patch provides basic Linux support for the bareflank
manager. It is capable of starting and stopping the
hypervisor given a list of modules to load.

This patch also contains the unit test for the bareflank
manager, as well as needed changes to the unit test
infrastructure to support more complicated unit testing.

Signed-off-by: Rian Quinn <quinnr@ainfosec.com>