New Build System
This patch contains a new build system. The VMM code actually
needs to be compiled twice. Once for the VMM itself using the
cross compiler, and once for the unit tests which need to
use the native compilers. The previous build system didn't
clearly define the difference.

This patch also contains the scafolding for a basic debugging
library for the VMM to demonstrate how to layout the different
files a class will need in the VMM

Signed-off-by: Rian Quinn <quinnr@ainfosec.com>