Work Towards VMCS Launch
There is a lot of work in this PR towards getting the VMCS to
launch. This includes:

- Most of the default VMCS initialization fields
- Most of the VMCS state debugging
- Most of the VMCS checks
- A basic exit handler
- LOTs of bug fixes and rework of the driver entry

Altough the launch is disabled, we have verified that we have a
working VMCS launch. There is still a decent amount of work to
get this working in master, as cleanup must be done prior to
offically enabling this feature.

Also, at the current moment we do not have an isolated exit
handler as it requires kernel resources to execute (i.e.
IDT, GDT, and CR3). This will be resolved in future versions
of the hypervisor.

Signed-off-by: Rian Quinn <quinnr@ainfosec.com>