New Memory Manager
The following provides a new memory manager for the VMM. With
this memory manager, it is now possible to new / delete in
the VMM. Since we do not yet have the MDL work complete, as
this will require std::array and std::hash, the VMM launch
code has been disabled for the moment. Once this code has
been complete, we will re-enable the code.

Signed-off-by: Rian Quinn <quinnr@ainfosec.com>