New Build System
The old build system has a lot of problems, as was very
inconsistent from one module to the next. This new one,
cleans up a lot of the issues that we had with the old system
and also addresses several bugs that we had in the past. It
also sets us up for being able to properly unittest all of the
modules.

Signed-off-by: Rian Quinn <quinnr@ainfosec.com>