VCPU Update
The following patch cleans up the VCPU code. During the
development of this patch, an issue came up with the way we
build the unit tests which was also cleaned up. Now, you
have to build the unit tests manually using make build_test
and these are done after all of the modules are built which
removes issues with cyclic dependencies

Signed-off-by: Rian Quinn <quinnr@ainfosec.com>