CR4 Shadow Bug
This patch fixes an issue with Linux when it flushed the TLB.
During the process of flusing the TLB, Linux writes some
CR4 bits, which cases Linux to read from a CR4 shadow, and
write the shadow bits to CR4. Since we were not updating CR4,
the VME bit was being cleared while VMXON was enabled, which
causes a GP. This patch updates the CR4 shadow to reflect
VME setting.

Signed-off-by: Rian Quinn <quinnr@ainfosec.com>