Fix Libcxx Permissions
There was a bug in the wrapper script that was basically allowing
certain flags to be passed to LD that should not have been. The
result was LD was thinkng that "-n" was being passed, which caused
the permissions of libc++.so to be RWE, which is not good once we
start turning on permissions. I also added relro and now, which
tells LD that we have no intention of lazy loading, and moved the
alignment setting to the script so that everything linked has
the same alignment (which actually speed up loading process a little)

Signed-off-by: Rian Quinn <quinnr@ainfosec.com>