DWARF Expression Suport
GCC 6.1 appears to add DWARF expressions for stack realignment
which is needed to support the addition of SSE instructions.
This patch provides the needed expression support for exceptions.
It should be noted that there appears to be a bug in GCC 6.1
that we have hacked around for now. For more information, please
see:

https://gcc.gnu.org/bugzilla/show_bug.cgi?id=71978

Signed-off-by: “Rian <“rianquinn@gmail.com”>