CPU-Z Crash (#143)
* Fix Windows 32bit App Crashing Issue

Windows 32bit applications were crashing because we were not
properly handling segment limits which are used by 32bit.
Specifically, when the granulatiry bit is set, the limit must
be converted to "bytes" for Intel VMX.

This patch also adds support for the LDTR which could be used
in the future. Just in case, the LDTR is now being set, but
usually is 0, even on Windows.

[ISSUE}: https://github.com/Bareflank/hypervisor/issues/140

Signed-off-by: “Rian <“rianquinn@gmail.com”>

* CPU-Z Crash

CPU-Z appears to be trying to access undefined MSRs. This patch
provides a "quick" in the RDMSR emulation code to handle these
undefined MSRs by ignoring them. It's possible that this might
need to be updated if more MSRs are found in the future on
other hardware

[ISSUE]: https://github.com/Bareflank/hypervisor/issues/142

Signed-off-by: “Rian <“rianquinn@gmail.com”>