VCPU Update
The previous VCPU implementation was put in place for 1.0
because we didn't have support for "gs" based state save
area. Now that we have that, we no longer need to call the
vcpu manager on every VM exit.

In addition, for version 1.2, we will need the ability to
create vcpuids that are no longer linear, and thus, an
array and hard coded MAX will not work. This patch removes
these limitations, and provides a means to have as many
vCPUs as needed in the future.

Signed-off-by: “Rian <“rianquinn@gmail.com”>