Destructor / Constructor Bug
The VCPU manager had an issue were the mutex lock could
deadlock if you printed from a constructor destructor. This
patch removes this issue.

The debug path was also not guarded from exceptions, which
would recursively hit the BS core. Added noexcept to
redirect exceptions to std::terminate if this should occur.

[ISUE]: https://github.com/Bareflank/hypervisor/issues/153

Signed-off-by: “Rian <“rianquinn@gmail.com”>