Add JSON Support
To support generic configuration files, and commands to the
hypervisor, we will support JSON. A generic VMCall will also
be supported, but unlikely to be used based on how easy
JSON is.

To support JSON, we are using a modified version of:
https://github.com/nlohmann/json

This is a really well tested JSON, header only implementation.
In addition, this implementation already supports native
integer types. Gutting float / double from this implementation
was trivial, meaning it will be easy to support newer versions
in the future.

To verify this works, double / float was gutted, and all of the
tests on the webstie were tested inside the VMM, and they all
worked great. If you attempt to use a floating point, the
compiler will detect this and error out.

Future PRs will add support for start_vmm, create_vcpu,
make_vcpu, and vmcall to support JSON, allowing extensions
to customize how this logic works. We will also provide
additional interfaces on BFM and the drivers to support
this.

Signed-off-by: “Rian <“rianquinn@gmail.com”>