Remove XSAVE (#151)
In order to support -O3, we need to save the YMM registers as
the hypervisor will use these are part of the optimizations.
The problem is, XSAVE has some issues with the hypervisor that
are not simple to solve.

For example, before you can run XSAVE, you need to set xcr0 with
the bits you plan to support. Since Barefalnk is both a bare
metal, and late launch hypervisor, the host OS is responsible
for setting up xcr0, which might not set the same bits that
Bareflank needs to use. This means Bareflank must enable the
bits it plans to use. If the host OS ever changes these
Bareflank would stop working. To solve that problem requires
a bit of emulation.

XSAVE is also saving a bunch of stuff that we don't need,
increasing the time it takes to Exit / Enter.

To solve these problems (and more), the simple solution is
to save off the contents of the YMM registers manually.
This patch provides the changes to do this.

In addition, this patch also changed how promote works.
It was amazing that the original code worked in "C", but
once optimizations were enabled, it stopped working.
In some cases, adding volatile helped (for example,
with GCC 6.1), but with the rest of the compilers, it was
still broken. This patch moved the promote code to
assembly so that these instructions can be handled carefully

Signed-off-by: “Rian <“rianquinn@gmail.com”>