Add Missing DWARF4 CFA Opcodes (#157)
We never implemented the opcodes in this patch, and they are
showing up during testing. The main reason we didn't implement
these was because it was assumed that GCC would not implement
stack operations in the EH Frame FDEs because they use a
malloc/free, which doesn't make sense if the throw came
from std::bad_alloc.

To overcome this limitation of malloc/free, we implemented the
stack frame "remember/restore" calls using a temp stack without
malloc/free. The down side to this is if the temp stack is not
large enough, the unwinder will fail, but the possibility of
this is extremely small, and even if it does occur, I added a
macro that can be increased if needed.

One thing that is interesting is it appears GCC is spitting
out bad DRAWF instructions for the case that was seen.
GCC is calling DW_CFA_restore on registers that do not have
initial instructions in the CIE, which results in 0.
The test cases still work because GCC is wrapping the bad
calls to DW_CFA_restore in a DW_CFA_remember_state /
DW_CFA_restore_state so in the end, the DW_CFA_restore are
simply ignored. Either way, this patch completes the missing
DWARF CFA instructons so we now implement them all.

Finally, we addressed an earlier issue with the expression
support that we thought was an issue with GCC but turned
out to be an issue with the unwinder. We were using
"<" instead of "<=" so we missed the last opcode which was
needed.

Signed-off-by: “Rian <“rianquinn@gmail.com”>