Update to Libc++ 3.9
This patch provides updates to support libc++ 3.9. It also
adds a patch queue for libc++. This will be used in the
future fix issues with libc++ that prevent certain features
from being used in the VMM (for example, unordered_map).
In addition, future PRs will provide a set of unit tests for
libc++ to verify that supported STL functionality works
as expected.

Signed-off-by: “Rian <“rianquinn@gmail.com”>