Remove delete[] from bfdrivers unit tests
This patch removes the need for delete[] in the bfdriver
unit tests. We already made this change in the unwinder tests
but never got around to it on the driver side. Interestingly
enough, this was the last remaining delete, everything else
at this point uses shared_ptr or unique_ptr.

[ISSUE]: https://github.com/Bareflank/hypervisor/issues/176

Signed-off-by: “Rian <“rianquinn@gmail.com”>