Add vmcs check control tests
This patch tests every function defined in
vmcs_intel_x64_check_controls.cpp.  The way in which the tests
are implemented is different from other modules.

Each function-under-test's control flow graph is represented
by a vector of lambdas that modify or enable/disable VMCS fields
and MSRs. Each element in the function's cfg represents exactly one
path from its control flow graph.  Each function-under-test corresponds
to exactly one member of vmcs_ut.  This member enumerates over the
cfg of the function-under-test, testing each element (i.e. path) in turn
with RUN_UNITTEST_WITH_MOCKS.

There are several advantages to testing this way.  First, all of the
logic to test a function f is contained in its corresponding
setup_f_paths function, rather than being spread out across several
setup functions that setup vmcs fields for many functions unrelated to
the one under test.  This simplifies verification and debugging of the
test itself because you only have to look at one function to see how it
works.

Also, as the underlying source changes, the test for that source need
only change in one place.  Implementors of tests don't have to worry
about the flip of a bit in one VMCS field breaking other unit tests
for unrelated functions.  The only data dependencies that need to be
tracked down in the event of a change are right there in one setup
function.  This aids comprehensibility and maintainability of the tests.