Thread Local Storage Fix
This PR re-enables thread local storage using the stack to store
the base address of the thread data. Libbfc has also been updated
to support this PR.

The only other issue that still needs to be addressed with respect
to multi-core is re-enabling thread statics, which might be an
issue if a static variable is defined in the exit handler.

[ISSUE]: https://github.com/Bareflank/hypervisor/issues/213

Signed-off-by: “Rian <“rianquinn@gmail.com”>