Initial Intrinsics Cleanup
This work is not even close to being completed. The goal of this PR
is to get this started so that many more PRs to follow can complete
this work entirely. While we are doing this, the repo will likely
be in a hybrid state, some code using the intrinsics class, and
some not in favor of the nested namespace approach.

For reference on what is being done here, please see the following:
https://github.com/Bareflank/hypervisor/issues/241

Signed-off-by: “Rian <“rianquinn@gmail.com”>