Revise vmlaunch tests
This commit creates setup functions for the vmcs check functions
that call solely check functions. This allows _every_ check function
to be tested in "isolation" or among others (i.e. as the root node
in the call graph or as a child node in the call graph).

The code uses the hippomocks "After" function to implement full coverage
for the failure case of vm::launch.  Before this PR, the write_*_state and
*_controls functions in vmcs_intel_x64::launch were overwriting
the vmcs fields after the setup functions had run.  Using "After",
test_vmlaunch_failure ensures that the setup functions are the only
ones that write to the vmcs state before the launch fails.

The rest of the setup functions for the guest checks will be added
once the other guest checks are merged in a later PR.