Remove Double Compile
This cleans up the dependencies of libcxx and libcxxabi, and
moves the compilation of bfcrt and bfunwind to after libcxx.
This not only cleans up a lot of the scripts and hackish
crap in the repo, but it also removes the double compile
problems. You can now:

./configure
make
make // <-- This should not recompile anything

Signed-off-by: “Rian <“rianquinn@gmail.com”>