Remove overuse of shared_ptr
We are using shared_ptr too much when it's really not needed.
This not only adds overhead, but it is misleading. This patch
removes all but a couple instances of shared_ptr where it is
not needed. A future patch will remove the remaining uses
in an attempt to keep this patch smaller.

Signed-off-by: “Rian <“rianquinn@gmail.com”>