Remove remaining #defines
This patch removes the remaining #defines that are not needed.
There is still a lot of cleanup with respect to aliases and
documentation, but this at least gets rid of the #defines
which is a good first start. Not that this is only for the
VMM. Since other code uses C, there are #defines were C is
used.

Signed-off-by: “Rian <“rianquinn@gmail.com”>