GSL Update
This updates the GSL to the latest and greatest, and cleans up our
patches to provide a more clear picture as to what we are
changing.

For whatever reason, updating the GSL causes GCC 6.2 to flag
several new bugs in the code that prevented the code from
compiling as well as passing the unit tests, so I also cleaned
up these newly discovered issues.

Signed-off-by: “Rian <“rianquinn@gmail.com”>