Deprecate GCC 5
This patch set removes GCC 5 and defaults to Clang 3.8 as the
cross compiler. In addition, newlib has been updated, and is now
linked in as a shared library instead of a static library.
Finally, when Clang is used, only Clang is used, unlike
before were GCC was used to compile Newlib

[ISSUE]: https://github.com/Bareflank/hypervisor/issues/336

Signed-off-by: “Rian <“rianquinn@gmail.com”>