AppVeyor Cleanup
This patch fixes a lot of issues with AppVeyor. Specifically, it doesn't
attempt to compile the VMM as that is being tested by Travis CI, and is
not specific to Windows. Instead, it focuses on the Windows specific
components being the driver and BFM. This also cleans up the setup
script for cygwin

Signed-off-by: “Rian <“rianquinn@gmail.com”>