Cache Flush / PAT Mapping / Bug Fixes
This PR adds cache flush to the repo, it also adds support for
the PATs. Up until this point, the hypevisor has been executing
with uncached memory, now it uses WB memory, and the maps can
be setup to use any memory type you want. If you map in guest
memory, it will match the guest mappings to prevent
inconsistent aliases.

This patch also removed a couple of bugs with the vCPU, and
Hippomocks, and added a mcro to silence verbosity with VMWare.

Signed-off-by: “Rian <“rianquinn@gmail.com”>