Add JSON Helpers
Since JSON doesn't support hex, we agreed in a previous RFC that
HEX would be represented using strings, which is fine since
everything starts out in string format anyways. To this end,
we need a simple way to get either a dec, or a hex value
for values and arrays. These helpers do that.

Signed-off-by: “Rian <“rianquinn@gmail.com”>