Serial Bug
The predefined port #s don't work in the serial namespace becasue
we cannot define all possible port #s. This became apparent when
trying to use the PCI based serial ports that were not predefined.
This patch gets rid of the predefine port numbers, allowing any
port number to be used.

Signed-off-by: “Rian <“rianquinn@gmail.com”>