Bug Fixes
This patch fixes too issues with the bfdriver. The first issue
is we were not verifying that the VMM was running when doing
a vmcall which causes all sorts of nasty things to happen if you
vmcall at the wrong time.

The second issue this addresses is we were not mapping in the
TLS data for all of the cores. We were only mapping in the
TLS for core 0. Not to mention we were never checking the
error case for the map in the first place, so this code was
just wrong on so many levels.

[ISSUE]: https://github.com/Bareflank/hypervisor/issues/394
[ISSUE]: https://github.com/Bareflank/hypervisor/issues/392

Signed-off-by: “Rian <“rianquinn@gmail.com”>