VMLauch Mods
This patch adds a traditional vmlaunch which is needed to start
guests with the VMCS. Now, you can declare a guest state as a
guest and not a host (the default), and vmlauch will be executed
without the demotion. This means you will have to specify
RIP and RSP manually until we add support for RIP and RSP in
the state class.

Signed-off-by: “Rian <“rianquinn@gmail.com”>