VMLauch Mods
This patch adds a traditional vmlaunch which is needed to start
guests with the VMCS. Now, you can declare a guest state as a
guest and not a host (the default), and vmlauch will be executed
without the demotion.

To support this, the vCPU has been relaxed so that it can be
not only be started, but it can also be resumed (i.e. run twice).
We also provided RIP / RSP fields in the state class so that
these registers can be defined for guest support.

Finally the VMCall completetion code has been moved so that
a vmcall can be completed from a different VM Exit.

Signed-off-by: “Rian <“rianquinn@gmail.com”>