ELF Loader Bug Fixes
The ELF loader was not working with ld.gold or ld.lld because these
linkers craete a rela.plt and use the old .ctors / .dtors sections.
They also do not properly label the eh_frame section. This patch
uses an updated ELF loader that is actually slated for 1.2, but has
these bug fixes in it.

Signed-off-by: Rian Quinn <“rianquinn@gmail.com”>