Remove Cross Compilers / ELF Loader Update
This patch removes the need for the cross compilers. You can now
install clang using your distro's package manager, and the scripts
will tell clang to cross compile, linking with the linker of your
choice, or defaulting to "ld". Note that the default clang that
is expected is clang 3.8, but you can set the compiler to whatever
you want, including GCC if you create the cross compiler youself.

The ELF loader was not working with ld.gold or ld.lld because these
linkers craete a rela.plt and use the old .ctors / .dtors sections.
They also do not properly label the eh_frame section. This patch
uses an updated ELF loader that is actually slated for 1.2, but has
these bug fixes in it.

Signed-off-by: Rian Quinn <“rianquinn@gmail.com”>