Map guest GDT read-write before promotion
This patch maps the guest's GDT into the hypervisor read/write
in exit_handler_intel_x64::handle_vmxoff.  This allows the
processor to set the TSS busy bit without it causing
a page fault if the guest has mapped in the GDT read-only
(Linux now exposes a read-only GDT, see 69218e4 and 45fc875
in the linux tree for more details).