Add bfarch.h to detect architecture and set flags
This header tests for predefined macros on both supported compilers to
determine the architecture being compiled for. Once detected, it defines
two macros:

    - BF_ARCH       string containing architecture name
    - BF_{archname} empty macro for #ifdef tests

Currently supported combinations are:

    - x64 on MSVC (via _MSC_VER and _X64)
    - x64 on GCC and clang (via __GNUC__, __clang__, and __x86_64__)
    - aarch64 on GCC and clang (via __GNUC__, __clang__, and __aarch64__)

Additionally, this header allows the user or the build system to modify
the architecture externally: defining BF_ARCH with -D disables the
checks and allows you to force an architecture flag.

Signed-off-by: Chris Pavlina <pavlinac@ainfosec.com>