Restructure bfvmm source and test files
This commit implements a basic restructuring of src and test files
within the bfvmm subtree, both to clean up build scripts for unit tests
and to prepare for multi-architecture support. The following is
implemented:

- src/foo/src/*.cpp (x64-related) -> src/foo/arch/intel_x64/*.cpp
- src/foo/src/*.cpp (non-specific) -> src/foo/*.cpp
- src/foo/test/*.cpp -> test/foo/*.cpp

No restructure of include files has been done yet - this is pending
discussion on related code restructure (namespacing, etc). A similar
split between architecture-specific and non-specific test code will be
done in the near future similar to the organization of the src files. A
similar restructuring of other subprojects is also planned for the near
future.

Signed-off-by: Chris Pavlina <pavlinac@ainfosec.com>