Fix Deadlock Issue
If the VM got a VMExit at the same time that it was trying to exit out
of the init process, deadlock would occur as printing to the console was
happening twice on the same CPU. This patch fixes this by moving the
console feedback to the CPUID instruction. v1.1 did something similar
using VMCall, but using CPUID prevents issues with VMCall casuing the
driver to explode when the hypervisor has not started properly, and
removes the need for the base hypervisor to implement VMcall.

[ISSUE]: https://github.com/Bareflank/hypervisor/issues/640

Signed-off-by: Rian Quinn <rianquinn@gmail.com>