Add intrinsics for EAPIs phys_mtrr class
Bit manipulation
================
The phys_mtrr class in the extended APIs needs to use
bsf and popcnt to quickly determine if the variable range
MTRRs overlap a given address.

- Add _bsf and _popcnt intrinsics.
- Add unit test stubs for _bsf and _popcnt.
- Add an overload to x64::is_physical_address_valid that takes
  the size of a physical address as an argument, as opposed to
  reading straight from cpuid.

Variable-range MTRRs
====================
There are a variable number of variable ranges supported by
a given system, as reported by the IA32_MTRRCAP msr. Because
of the dynamic number of ranges, it is not useful to have
separate physbase{n}, phymask{n} namespaces for each variable
MTRR that is supported.

The fields within each variable MTRR have the same format, so it
is much cleaner and simpler to have one ia32_physbase and
ia32_physmask namespace for all of them.

- Remove the physbase[0-9], physmask[0-9] namespaces.
- Add single ia32_physbase and ia32_physmask namespaces for parsing
  the variable-range MTRRs.

Fixed-range MTRRs
=================
Each fixed-range MTRR encodes the type of eight subranges with
one type per byte. So far the eapis code has not required statically
parsing the range types with the range[0-7] subnamespaces.

- Remove the range[0-7] sub-namespaces of each fixed-range MTRR
  namespace.
- Add a single 'namespace fixed_range' that contains a 'type' function
  that returns the type of the subrange requested.

Signed-off-by: Connor Davis <davisc@ainfosec.com>