Fix for #702 [BUG]: VMM does not handle NMIs
Initialize the host IDT with a default exception handler and an NMI handler.
Note that interrupt handlers are not installed, as those are taken care
of in the extended apis.

  - The non-NMI handlers print out the CPU state at the time of the
    exception and halt the CPU.

  - The NMI handler enables NMI-window exiting. This ensures that an NMI
    can occur on any exit, anywhere the VMM has an IDT, and the NMI will
    be injected as soon as the window opens in the guest.

VT-x requires NMI-exiting and virtuall-NMIs controls to be set when
NMI-window exiting is set, so two exit handlers are added; one for
NMI exits, the other for NMI-window exits. The base exit handler
registers these during construction.

Add unit tests for new code.

Signed-off-by: Connor Davis <davisc@ainfosec.com>