RSDP and Init/Fini handlers
This patch adds the ability for the drvier to report the location
of the RSDP, which is far easier to get from the OS than it is
to search for since it can be relocated. The use of this is
optional, and it should be assumed that the driver could report
a nullptr, and that doesn't suggest ACPI doesn't exist.

This also adds a set of init/fini handlers. The init handler is
executed after a launch and allows the VMM to initialize itself
further after it has started. The fini handler is executed just
before completing.

[ISSUE]: https://github.com/Bareflank/hypervisor/issues/718