New End User Interface
This patch provides a new end user interface for creating extensions
that is more "C" like in nature. The big change that was needed to
support this patch was the removal of the gsl::not_null from delegate
types which is not needed since the APIs can ensure that the vcpus are
always valid making the APIs more accessible to C programmers.

We also removed dead code in the entry.cpp file that was added when the
RSDP code was added but never removed, and we added two additional
init functions for the new C interface