Use VMM toolchain config to build newlib
Prior to this change, newlib was always built using the system clang
binary. Since newlib is a VMM dependency it should be compiled and
linked using the compiler and linker specified in the VMM toolchain
file. Other dependencies such as libcxxabi are built using the
VMM toolchain file as seen in libcxxabi.cmake.

[ISSUE]: https://github.com/Bareflank/hypervisor/issues/753

Signed-off-by: Matt Dees <mdees@uci.edu>