cmake: Fix bfsdk, LOG_BUILD bugs (#850)
Remove the bfdso dependency from libcxx. The dso object is built as a
runtime component to which libcxxabi links in the VMM's final link
step. Removing the subproject dependency ensures that libcxxabi and
libcxx are not checked for changes on every make

Add top-level CMakeLists.txt for each example extension. Now each
example should be able to be built as an external extension, even though
they reside in-tree.

Pass explicit 'LOG_BUILD' to ExternalProject_Add (EPA) so that log files
are written to the stamp directory. Omitting this can cause EPA to
put log files in a non-existent directory in parallel builds.

Signed-off-by: Connor Davis <davisc@ainfosec.com>