Add initial scaffolding for the vmmctl and the Linux loader
This patch adds to initial scaffolding needed to use the vmmctl
to talk to a the Linux loader. This will provide the starting
point to implement the Linux loader. Once the VMM is able to
start/stop from the loader on Windows and Linux, we will finish
the vmmctl logic, including tests, Perforce, etc...