Adds additional loader scaffolding
This patch attempts to add in additional scaffolding so that
where all of the logic is located is easier to identify. To
ensure that everything works as needed, this code also adds
in the per-cpu logic that will be needed to launch on each
cpu. Some improvements were made to work without the need for
symbol lookups on Linux as well.