vcpu: Use aligned moves in exception handlers
The Intel manual recommends that interrupt stacks are aligned to
16-bytes so that aligned moves can be used for saving and restoring xmm
registers.

This patch ensures that aligned moves can be used in both the error-code
and non-error-code exception handlers by pushing an extra 8-byte value
in non-error-code handlers. It also rounds the value returned from
setup_stack to the next-lowest 16-byte boundary before placing it in the ist1
entry of the vcpu's host TSS structure.

With these two changes movdqa is used for saving/restoring the xmm
registers. In addition, the xmm8-xmm15 and mxcsr registers are also
saved/restored so that all SSE state is properly managed.