ept: Add invept API
This patch add an invept() function to the vcpu and ept_handler classes.
It also adds m_eptp and m_invept members to the ept_handler class. The
m_eptp member tracks the value of the eptp field of the vcpu's vmcs. The
m_invept member is a function pointer used to implement invalidation.

If single-context invalidation is supported, then m_invept points to
::intel_x64::vmx::invept_single_context and the m_eptp member is used as
the eptp argument (the value must match the vmcs field in its entirety,
not just the address portion). Otherwise the global version is
used.

The existing uses of ::intel_x64::vmx::invept_global in the
control_register class are replaced with calls to vcpu->invept() which
is more granular assuming single-context invalidation is supported.