vcpu: Fix xsave init bugs
Subleaf 0x1 of leaf 0xD reports the size of the xsave area needed
by the currently enabled features in (XCR0 | IA32_XSS). This patch
ensures that XCR0 and IA32_XSS contain the maximum number of bits set
as reported by CPUID before reading the size.

The old code was using the values without first setting them to their
maximum allowed value. However since the size returned is rounded up to
at least 4KB, it would never be a problem as long as the max xsave area
is less than 4KB as well.

This patch also fixes an issue where the xsave CPUID leaf was being read
before checking that the feature is implmented (as reported by leaf 1).